requirements.txt
pandas>=2.1.0
matplotlib>=3.8.0
networkx>=3.2.1
PyYAML>=6.0.1
psutil>=5.9.8
numpy>=1.26.4
